refactoring

+++ src/main/java/com/bank/bank_system/exception/InsufficientBalanceException.java
@@ -1,7 +1,9 @@
 package com.bank.bank_system.exception;
 
+import java.math.BigDecimal;
+
 public class InsufficientBalanceException extends RuntimeException {
-    public InsufficientBalanceException(Long accountId, Double amount) {
+    public InsufficientBalanceException(Long accountId, BigDecimal amount) {
         super("Insufficient balance in account " + accountId + " to withdraw " + amount);
     }
